[ticket/15905] Try with existing phpbb extension
PHPBB3-15905
