+++ tests/controller/common_helper_route.php
@@ -74,7 +74,9 @@ abstract class phpbb_controller_common_helper_route extends phpbb_test_case
 
 	protected function generate_route_objects()
 	{
-		global $request;
+		global $request, $phpbb_root_path, $phpEx;
+
+		$auth = $this->getMock('\phpbb\auth\auth');
 
 		$this->request = new phpbb_mock_request();
 		$this->request->overwrite('SCRIPT_NAME', $this->get_uri(), \phpbb\request\request_interface::SERVER);
@@ -122,7 +124,7 @@ abstract class phpbb_controller_common_helper_route extends phpbb_test_case
 				'autoescape'	=> false,
 			)
 		);
-		$this->template = new phpbb\template\twig\twig($this->phpbb_path_helper, $this->config, $context, $twig, $cache_path, $this->user, array(new \phpbb\template\twig\extension($context, $twig, $this->user)));
+		$this->template = new phpbb\template\twig\twig($this->phpbb_path_helper, $this->config, $context, $twig, $cache_path, $this->user, array(new \phpbb\template\twig\extension($auth, $context, $twig, $this->user)));
 		$twig->setLexer(new \phpbb\template\twig\lexer($twig));
 
 		$this->extension_manager = new phpbb_mock_extension_manager(
