+++ tests/email/email_parsing_test.php
@@ -66,7 +66,15 @@ class phpbb_email_parsing_test extends phpbb_test_case
 		);
 		$phpbb_container->set('ext.manager', $extension_manager);
 
+		$auth = $this->getMock('\phpbb\auth\auth');
 		$context = new \phpbb\template\context();
+		$twig_extension = new \phpbb\template\twig\extension($auth, $context, $lang);
+		$phpbb_container->set('template.twig.extensions.phpbb', $twig_extension);
+
+		$twig_extensions_collection = new \phpbb\di\service_collection($phpbb_container);
+		$twig_extensions_collection->add('template.twig.extensions.phpbb');
+		$phpbb_container->set('template.twig.extensions.collection', $twig_extensions_collection);
+
 		$twig = new \phpbb\template\twig\environment(
 			$config,
 			$filesystem,
